fixes spent states of addresses not getting persisted on tx pruning
If a node only boots up with a database (without spent-addresses-db) and the
node has LS pruning enabled, then the node will prune everything below
the given thresholds but not persist the spent state of the corresponding
addresses.

This commit fixes the issue by replacing the spentAddressesService's
wasAddressSpentFrom() with a simple check against the persistence layer
(spentAddressesProvider.containsAddress()) in order to ensure, that the
spent state is correctly persisted. Previously the check would always return
true, making it so that the address was never persisted as spent.

Signed-off-by: Luca Moser <moser.luca@gmail.com>